Replace the default osmdroid "you are here" yellow peg man with a bike
This requires adding a ResourceProxy implementation to intercept
requests for the "person" bitmap and substitute our own bitmap. Note
that our implementation can't inherit DefaultResourceProxy because
DefaultResourceProxy uses getClass() and thus pulls all resources from
the project the most derived class is in, not its own.

The bike png is not transparent and doesn't stand out enough, but that
can be replaced later.