Switch AboutActivity to fragments
For the first small step in moving to fragments, I've put the
AboutActivity layout into a new About fragment. AboutActivity is now
trivial with a trivial layout that just contains the fragment. Note that
I've added the "Android support library" so that fragments work on
pre-3.0 devices too.

There's no user-visible change yet.